refactor(drawer): move regularText CSS to ConfigOption
Throughout all uses of ConfigOption, the title variant and title
classnames follow a certain pattern. For mobile responsive fields,
they start with h6 variant and use regularText CSS when in mobile view

Given that this pattern has repeated for many times, move the
implementation into ConfigOption's implementation and expose a boolean
isMobileResponsive flag to toggle the behaviour.
